fix(daemon): host runtime contract validators in a dependency-free module
The contract closure gates run without node_modules, so files reachable
from daemon-protocol.contract.ts must stay dependency-free; the kernel
barrel pulls effect. Move DTO types, validators and serializers into
agent-runtime-contract.ts (no runtime imports) and keep the read model
importing kernel values through the public barrel, outside the closure.
