+++ packages/daemon/src/agent-runtime-contract.ts
@@ -1,0 +1,21 @@
+import type { AgentRuntimeEventV1 } from "../../kernel/src/index.ts";
+export interface AgentRuntimeInstallationDto { readonly installationId: string; readonly kindId: string; readonly protocolFamily: "claude-compatible" | "codex"; readonly version: string; readonly credentialState: "configured" | "not_configured" | "invalid"; readonly attachCapability: "supported" | "unsupported"; readonly lastObservedAt: string }
+export interface AgentRuntimeAssociationDto { readonly taskId: string; readonly executionId: string; readonly holder: { readonly personId: string; readonly executorId: string | null } | null; readonly lease: { readonly phase: "reserving" | "active" | "released" | "orphaned"; readonly expiresAt: string } | null }
+export interface AgentRuntimeSessionDto { readonly runtimeSessionId: string; readonly providerSessionId: string | null; readonly installationId: string; readonly kindId: string; readonly credentialState: AgentRuntimeInstallationDto["credentialState"]; readonly liveness: "live" | "stale" | "unknown" | "exited"; readonly attachCapability: "supported" | "unsupported"; readonly streamCursor: string; readonly associations: readonly AgentRuntimeAssociationDto[]; readonly activity: { readonly lastObservedAt: string; readonly outcome: "succeeded" | "failed" | "unknown" | null; readonly resultRef: string | null } }
+export interface AgentRuntimeLifecycleDto { readonly cursor: string; readonly runtimeSessionId: string; readonly type: AgentRuntimeEventV1["type"]; readonly occurredAt: string }
+export type AgentRuntimeOverviewResult = { readonly ok: true; readonly status: "ready" | "pending"; readonly installations: readonly AgentRuntimeInstallationDto[]; readonly sessions: readonly AgentRuntimeSessionDto[]; readonly watermark: number; readonly sourceRevision: number };
+export type AgentRuntimeSessionResult = { readonly ok: true; readonly status: "ready" | "pending"; readonly session: AgentRuntimeSessionDto; readonly watermark: number; readonly sourceRevision: number };
+export type AgentRuntimeEventsResult = { readonly ok: true; readonly runtimeSessionId: string; readonly events: readonly AgentRuntimeLifecycleDto[]; readonly cursor: string; readonly sourceCursor: string; readonly done: boolean };
+export function coded(code: string, message: string): Error { const error = new Error(message) as Error & { code: string }; error.code = code; return error; }
+const credentialStates = ["configured", "not_configured", "invalid"], liveness = ["live", "stale", "unknown", "exited"], attachCapabilities = ["supported", "unsupported"];
+export function validateAgentRuntimeOverview(value: unknown): readonly string[] { return record(value) && exact(value, ["ok", "status", "installations", "sessions", "watermark", "sourceRevision"]) && value.ok === true && ["ready", "pending"].includes(String(value.status)) && Array.isArray(value.installations) && value.installations.every(validInstallation) && Array.isArray(value.sessions) && value.sessions.every(validSession) && Number.isInteger(value.watermark) && Number.isInteger(value.sourceRevision) && safeKeys(value) ? [] : ["agent runtime overview is invalid"]; }
+export function validateAgentRuntimeSession(value: unknown): readonly string[] { return record(value) && exact(value, ["ok", "status", "session", "watermark", "sourceRevision"]) && value.ok === true && ["ready", "pending"].includes(String(value.status)) && validSession(value.session) && Number.isInteger(value.watermark) && Number.isInteger(value.sourceRevision) && safeKeys(value) ? [] : ["agent runtime session is invalid"]; }
+export function validateAgentRuntimeEvents(value: unknown): readonly string[] { return record(value) && exact(value, ["ok", "runtimeSessionId", "events", "cursor", "sourceCursor", "done"]) && value.ok === true && typeof value.runtimeSessionId === "string" && Array.isArray(value.events) && value.events.every((event) => record(event) && exact(event, ["cursor", "runtimeSessionId", "type", "occurredAt"]) && /^lifecycle:\d+$/u.test(String(event.cursor)) && event.runtimeSessionId === value.runtimeSessionId && typeof event.type === "string" && typeof event.occurredAt === "string") && /^lifecycle:\d+$/u.test(String(value.cursor)) && /^lifecycle:\d+$/u.test(String(value.sourceCursor)) && typeof value.done === "boolean" && safeKeys(value) ? [] : ["agent runtime lifecycle events are invalid"]; }
+function validInstallation(value: unknown): value is AgentRuntimeInstallationDto { return record(value) && exact(value, ["installationId", "kindId", "protocolFamily", "version", "credentialState", "attachCapability", "lastObservedAt"]) && typeof value.installationId === "string" && typeof value.kindId === "string" && ["claude-compatible", "codex"].includes(String(value.protocolFamily)) && typeof value.version === "string" && credentialStates.includes(String(value.credentialState)) && attachCapabilities.includes(String(value.attachCapability)) && typeof value.lastObservedAt === "string"; }
+function validSession(value: unknown): value is AgentRuntimeSessionDto { return record(value) && exact(value, ["runtimeSessionId", "providerSessionId", "installationId", "kindId", "credentialState", "liveness", "attachCapability", "streamCursor", "associations", "activity"]) && typeof value.runtimeSessionId === "string" && (value.providerSessionId === null || typeof value.providerSessionId === "string") && typeof value.installationId === "string" && typeof value.kindId === "string" && credentialStates.includes(String(value.credentialState)) && liveness.includes(String(value.liveness)) && attachCapabilities.includes(String(value.attachCapability)) && /^stream:\d+$/u.test(String(value.streamCursor)) && Array.isArray(value.associations) && value.associations.every(validAssociation) && record(value.activity) && exact(value.activity, ["lastObservedAt", "outcome", "resultRef"]) && typeof value.activity.lastObservedAt === "string" && (value.activity.outcome === null || ["succeeded", "failed", "unknown"].includes(String(value.activity.outcome))) && (value.activity.resultRef === null || typeof value.activity.resultRef === "string"); }
+function validAssociation(value: unknown): value is AgentRuntimeAssociationDto { return record(value) && exact(value, ["taskId", "executionId", "holder", "lease"]) && typeof value.taskId === "string" && typeof value.executionId === "string" && (value.holder === null || record(value.holder) && exact(value.holder, ["personId", "executorId"]) && typeof value.holder.personId === "string" && (value.holder.executorId === null || typeof value.holder.executorId === "string")) && (value.lease === null || record(value.lease) && exact(value.lease, ["phase", "expiresAt"]) && ["reserving", "active", "released", "orphaned"].includes(String(value.lease.phase)) && typeof value.lease.expiresAt === "string"); }
+function safeKeys(value: unknown): boolean { if (Array.isArray(value)) return value.every(safeKeys); if (!record(value)) return true; for (const [key, nested] of Object.entries(value)) { if (["credential", "credentialValue", "secret", "token", "environment", "argv", "transcript", "stdout", "stderr"].includes(key) || !safeKeys(nested)) return false; } return true; }
+export const serializeAgentRuntimeOverview = (value: unknown): string => serialize(value, validateAgentRuntimeOverview), serializeAgentRuntimeSession = (value: unknown): string => serialize(value, validateAgentRuntimeSession), serializeAgentRuntimeEvents = (value: unknown): string => serialize(value, validateAgentRuntimeEvents);
+function serialize(value: unknown, validate: (candidate: unknown) => readonly string[]): string { const errors = validate(value); if (errors.length) throw coded("invalid_result", errors.join("; ")); return `${JSON.stringify(value)}\n`; }
+function record(value: unknown): value is Record<string, unknown> { return value !== null && typeof value === "object" && !Array.isArray(value); }
+function exact(value: Record<string, unknown>, fields: readonly string[]): boolean { return Object.keys(value).length === fields.length && fields.every((field) => Object.hasOwn(value, field)); }
