+++ packages/daemon/src/protocol/daemon-protocol.contract.ts
@@ -1,7 +1,7 @@
 import { existsSync, realpathSync } from "node:fs";
 import path from "node:path";
 import { isJsonObject, type JsonObject } from "./json-rpc-types.ts";
-import { validateAgentRuntimeEvents, validateAgentRuntimeOverview, validateAgentRuntimeSession, type AgentRuntimeEventsResult, type AgentRuntimeOverviewResult, type AgentRuntimeSessionResult } from "../agent-runtime-read.ts";
+import { validateAgentRuntimeEvents, validateAgentRuntimeOverview, validateAgentRuntimeSession, type AgentRuntimeEventsResult, type AgentRuntimeOverviewResult, type AgentRuntimeSessionResult } from "../agent-runtime-contract.ts";
 import { validateAgentRuntimeAttach, validateAgentRuntimeAttachEvent, type AgentRuntimeAttachEvent, type AgentRuntimeAttachResult } from "../agent-runtime-stream.ts";
 declare const safePathBrand: unique symbol, canonicalRootBrand: unique symbol, endpointIdentityBrand: unique symbol, workspaceIdBrand: unique symbol;
 export type SafePath = string & { readonly [safePathBrand]: true }; export type CanonicalRoot = SafePath & { readonly [canonicalRootBrand]: true };
@@ -82,8 +82,8 @@ export const daemonGuiReadSchemas = Object.freeze([{ id: DAEMON_TASK_SNAPSHOT_LI
   { id: DAEMON_RELATION_GRAPH_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_RELATION_GRAPH_SCHEMA", parser: "packages/daemon/src/protocol/daemon-protocol.contract.ts#validateDaemonRelationGraph", writer: "packages/daemon/src/protocol/daemon-protocol.contract.ts#serializeDaemonRelationGraph", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-relation-graph-invalid.json"]) },
   { id: DAEMON_DECISION_LIST_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_DECISION_LIST_SCHEMA", parser: "packages/daemon/src/protocol/daemon-protocol.contract.ts#validateDaemonDecisionList", writer: "packages/daemon/src/protocol/daemon-protocol.contract.ts#serializeDaemonDecisionList", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-decision-list-invalid.json"]) },
   { id: DAEMON_DOCUMENT_READ_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_DOCUMENT_READ_SCHEMA", parser: "packages/daemon/src/protocol/daemon-protocol.contract.ts#validateDaemonDocumentRead", writer: "packages/daemon/src/protocol/daemon-protocol.contract.ts#serializeDaemonDocumentRead", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-document-read-invalid.json"]) }, { id: DAEMON_PROTOCOL_ERROR_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_PROTOCOL_ERROR_SCHEMA", parser: "packages/daemon/src/protocol/daemon-protocol.contract.ts#validateDaemonProtocolError", writer: "packages/daemon/src/protocol/daemon-protocol.contract.ts#serializeDaemonProtocolError", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-protocol-error-invalid.json"]) },
-  { id: DAEMON_AGENT_RUNTIME_OVERVIEW_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_AGENT_RUNTIME_OVERVIEW_SCHEMA", parser: "packages/daemon/src/agent-runtime-read.ts#validateAgentRuntimeOverview", writer: "packages/daemon/src/agent-runtime-read.ts#serializeAgentRuntimeOverview", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-agent-runtime-overview-invalid.json"]) }, { id: DAEMON_AGENT_RUNTIME_SESSION_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_AGENT_RUNTIME_SESSION_SCHEMA", parser: "packages/daemon/src/agent-runtime-read.ts#validateAgentRuntimeSession", writer: "packages/daemon/src/agent-runtime-read.ts#serializeAgentRuntimeSession", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-agent-runtime-session-invalid.json"]) },
-  { id: DAEMON_AGENT_RUNTIME_EVENTS_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_AGENT_RUNTIME_EVENTS_SCHEMA", parser: "packages/daemon/src/agent-runtime-read.ts#validateAgentRuntimeEvents", writer: "packages/daemon/src/agent-runtime-read.ts#serializeAgentRuntimeEvents", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-agent-runtime-events-invalid.json"]) }, { id: DAEMON_AGENT_RUNTIME_ATTACH_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_AGENT_RUNTIME_ATTACH_SCHEMA", parser: "packages/daemon/src/agent-runtime-stream.ts#validateAgentRuntimeAttach", writer: "packages/daemon/src/agent-runtime-stream.ts#serializeAgentRuntimeAttach", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-agent-runtime-attach-invalid.json"]) }, { id: DAEMON_AGENT_RUNTIME_ATTACH_EVENT_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_AGENT_RUNTIME_ATTACH_EVENT_SCHEMA", parser: "packages/daemon/src/agent-runtime-stream.ts#validateAgentRuntimeAttachEvent", writer: "packages/daemon/src/agent-runtime-stream.ts#serializeAgentRuntimeAttachEvent", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-agent-runtime-attach-event-invalid.json"]) }]);
+  { id: DAEMON_AGENT_RUNTIME_OVERVIEW_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_AGENT_RUNTIME_OVERVIEW_SCHEMA", parser: "packages/daemon/src/agent-runtime-contract.ts#validateAgentRuntimeOverview", writer: "packages/daemon/src/agent-runtime-contract.ts#serializeAgentRuntimeOverview", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-agent-runtime-overview-invalid.json"]) }, { id: DAEMON_AGENT_RUNTIME_SESSION_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_AGENT_RUNTIME_SESSION_SCHEMA", parser: "packages/daemon/src/agent-runtime-contract.ts#validateAgentRuntimeSession", writer: "packages/daemon/src/agent-runtime-contract.ts#serializeAgentRuntimeSession", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-agent-runtime-session-invalid.json"]) },
+  { id: DAEMON_AGENT_RUNTIME_EVENTS_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_AGENT_RUNTIME_EVENTS_SCHEMA", parser: "packages/daemon/src/agent-runtime-contract.ts#validateAgentRuntimeEvents", writer: "packages/daemon/src/agent-runtime-contract.ts#serializeAgentRuntimeEvents", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-agent-runtime-events-invalid.json"]) }, { id: DAEMON_AGENT_RUNTIME_ATTACH_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_AGENT_RUNTIME_ATTACH_SCHEMA", parser: "packages/daemon/src/agent-runtime-stream.ts#validateAgentRuntimeAttach", writer: "packages/daemon/src/agent-runtime-stream.ts#serializeAgentRuntimeAttach", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-agent-runtime-attach-invalid.json"]) }, { id: DAEMON_AGENT_RUNTIME_ATTACH_EVENT_SCHEMA.id, schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_AGENT_RUNTIME_ATTACH_EVENT_SCHEMA", parser: "packages/daemon/src/agent-runtime-stream.ts#validateAgentRuntimeAttachEvent", writer: "packages/daemon/src/agent-runtime-stream.ts#serializeAgentRuntimeAttachEvent", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError", negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/daemon-agent-runtime-attach-event-invalid.json"]) }]);
 export default Object.freeze({ id: "w3-daemon-protocol", phases: Object.freeze(["W3", "W2-GUI", "DocSync-B", "Runtime-B"]), commands: daemonProtocolCommands, methods: allDaemonProtocolMethods, gates: Object.freeze([]), guards: Object.freeze([]), schemas: Object.freeze([{ id: DAEMON_RPC_SCHEMA.id,
     schema: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DAEMON_RPC_SCHEMA", parser: "packages/daemon/src/protocol/daemon-protocol.contract.ts#validateDaemonRpcCall", writer: "packages/daemon/src/protocol/daemon-protocol.contract.ts#serializeDaemonRpcCall", error: "packages/daemon/src/protocol/daemon-protocol.contract.ts#DaemonProtocolContractError",
     negativeFixtures: Object.freeze(["packages/daemon/fixtures/contracts/w3-daemon-rpc-invalid.json"]) }, ...daemonGuiReadSchemas]) });
